Projeto utilizado no curso React: gerenciando estado com Recoil
