+++ javascript/react-gerenciado-estado-com-recoil/db.json
@@ -1,0 +1,32 @@
+{
+  "eventos": [
+    {
+      "descricao": "Estudar React",
+      "inicio": "2022-01-15T09:00",
+      "fim": "2022-01-15T13:00",
+      "completo": false,
+      "id": 101
+    },
+    {
+      "descricao": "Estudar TypeScript",
+      "inicio": "2022-01-14T09:00",
+      "fim": "2022-01-14T13:00",
+      "completo": false,
+      "id": 102
+    },
+    {
+      "descricao": "Estudar Padrões de Projeto",
+      "inicio": "2022-01-13T09:00",
+      "fim": "2022-01-13T13:00",
+      "completo": false,
+      "id": 103
+    },
+    {
+      "descricao": "Estudar Recoil",
+      "inicio": "2022-01-16T09:00",
+      "fim": "2022-01-16T11:00",
+      "completo": false,
+      "id": 104
+    }
+  ]
+}
