+++ javascript/react-gerenciado-estado-com-recoil/public/manifest.json
