+++ javascript/react-gerenciado-estado-com-recoil/src/components/Card/Card.module.scss
