+++ javascript/react-gerenciado-estado-com-recoil/src/components/Evento/index.tsx
