+++ javascript/react-gerenciado-estado-com-recoil/src/components/ListaDeEventos/index.tsx
