+++ javascript/react-gerenciado-estado-com-recoil/src/interfaces/IEvento.ts
