+++ javascript/react-gerenciado-estado-com-recoil/src/interfaces/IFiltroDeEventos.ts
