+++ javascript/react-gerenciado-estado-com-recoil/src/state/atom.ts
@@ -1,0 +1,14 @@
+import { atom } from "recoil";
+import { IEvento } from "../interfaces/IEvento";
+import { IFiltroDeEventos } from "../interfaces/IFiltroDeEventos";
+import { eventosAsync } from "./seletores";
+
+export const listaDeEventosState = atom<IEvento[]>({
+  key: 'listaDeEventosState',
+  default: eventosAsync
+})
+
+export const filtroDeEventos = atom<IFiltroDeEventos>({
+  key: 'filtroDeEventos',
+  default: {}
+})
