+++ javascript/react-gerenciado-estado-com-recoil/src/state/hooks/useListaDeEventos.ts
