+++ javascript/react-gerenciado-estado-com-recoil/src/state/seletores/index.ts
@@ -1,4 +1,5 @@
 import { selector } from "recoil";
+import { IEvento } from "../../interfaces/IEvento";
 import { filtroDeEventos, listaDeEventosState } from "../atom";
 
 export const eventosFiltradosState = selector({
@@ -17,3 +18,15 @@ export const eventosFiltradosState = selector({
   }
 })
 
+export const eventosAsync = selector({
+  key: 'eventosAsync',
+  get: async () => {
+    const respostaHttp = await fetch('http://localhost:8080/eventos')
+    const eventosJson: IEvento[] = await respostaHttp.json()
+    return eventosJson.map(evento => ({
+      ...evento,
+      inicio: new Date(evento.inicio),
+      fim: new Date(evento.fim)
+    }))
+  }
+})
