+++ javascript/react-gerenciado-estado-com-recoil/tsconfig.json
@@ -19,7 +19,8 @@
     "isolatedModules": true,
     "noEmit": true,
     "jsx": "react-jsx",
-    "plugins": [{ "name": "typescript-plugin-css-modules" }]
+    "plugins": [{ "name": "typescript-plugin-css-modules" }],
+    "downlevelIteration": true
   },
   "include": [
     "src"
